==> mount/preprocess/generate_single_cell_data/0.generate_pseudobulk_mouse.Rmd <==

# reading in mouse pseudobulk Rds files

```{r load_mouse_rds}
if(!file.exists("alldata_mouse.Rds")){
  rds = list.files(path = '../mmus_samples_panglaodb/', pattern = '.Rds.gz', recursive = T)
  raw_alldata = sapply(paste('../mmus_samples_panglaodb/', rds, sep = ''), function(x) { 
    y <- gzfile(x)
    z <- gzcon(y)
    data = readRDS(z)
    close(z)
    data })
}
```

# relabing mouse clusters

```{r relabel_mouse_rds}
if ('raw_alldata' %in% ls()) {
fixed = sapply(1:length(raw_alldata),function(x) {
  srs_names = sapply(sapply(basename(rds[x]), strsplit, '_'), function(x) grep('^SR',x,value=T) ,simplify = F)
  print(srs_names)
  tissue = attributes(raw_alldata[[x]])$tissue
  newids = paste(srs_names[[1]][2],tissue,dimnames(raw_alldata[[x]])[[1]],sep=':')
  newids = paste(ifelse(attributes(raw_alldata[[x]])$clustersize < 200, 'SMALL_','' ),newids,sep='')
  dimnames(raw_alldata[[x]]) <- list(newids, dimnames(raw_alldata[[x]])[[2]])
  attributes(raw_alldata[[x]])$sizes <- setNames(attributes(raw_alldata[[x]])$clustersize,rownames(raw_alldata[[x]]))
  raw_alldata[[x]]
})
}
```

# concatenating Rds files

```{r concat_mouse_rds}
if ('fixed' %in% ls()) {
  alldata = do.call(function(...) {
    tmp <- plyr::rbind.fill.matrix(...)
    cols = dimnames(tmp)[[2]]
    rows <- sapply(fixed, function(i) {
      dimnames(i)[[1]]
    })
    dimnames(tmp) = list(unlist(rows),cols)
    return(tmp)

  }, fixed)
  sizes = do.call(c,lapply(fixed,function(set){ attributes(set)$sizes }))
  attributes(alldata)$clustersize = sizes
  rm(fixed)
  saveRDS(alldata,'alldata_mouse.Rds')
  rm(alldata)
}
```
