==> mount/preprocess/generate_single_cell_data/1.generate_pseudobulkpresence.Rmd <==

# loading pseudopresence model

```{r pseudopresence_model1}
pseudopresence_model.lm <- readRDS('lm_glycogene_pseudopresence.Rds')
compute_sample_pop <- function(pseudobulk, mu_model.lm, sd_model.lm) {

  mu <-
    stats::predict(
      mu_model.lm,
      newdata = data.frame(pseudobulk_mean = pseudobulk),
      se.fit = TRUE,
      level = 0.95,
      interval = "confidence")$fit[,'fit']

  s <-
    stats::predict(
      sd_model.lm,
      newdata = data.frame(pseudobulk = pseudobulk),
      se.fit = TRUE,
      level = 0.95,
      interval = "confidence")$fit[,'fit']

  return(list(mu = mu, s = s))

}
```

```{r pseudopresence_model2}
## computing pvalue for each glycogene on each cluster
compute_pseudopresence_pvalue <- function(pseudobulk, mu, s, n) {

  # n chosen based on the number of pseudobulks that model is fit to
  ts  <- (mu - pseudobulk)/(s/sqrt(n))
  p   <- 2*pt(-abs(ts), n-1)

  return(p)

}
```

# loading pseudobulk model

```{r pseudobulk_model1}
expression_model.lm <- readRDS('lm_chromium10x_mupb_norm_nziqr_threshold_cutoff.Rds')
cutoff.int <- mean(expression_model.lm$xy.df$y[1:7])

compute_threshold <- function(ncells) {

  s <-
    stats::predict(
      expression_model.lm$model2.lm,
      newdata = data.frame(x = ncells),
      se.fit = TRUE,
      level = 0.95,
      interval = "confidence")$fit[,'fit']

  return(s)

}
```

```{r pseudobulk_model2}
compute_pvalue <- function(pseudobulk, threshold) {

  n   <- 7 # chosen based on the number of pseudobulk sizes that model is fit to
  ts  <- (cutoff.int - pseudobulk)/(threshold/sqrt(n))
  p   <- 2*pt(-abs(ts), n-1)

  return(p)

}

expression_status_call <- function(threshold,pseudobulk) {
  if (is.na(pseudobulk)) {
    return("unexpressed")
  }
  if (pseudobulk == 0) {
    return("unexpressed")
  }

  expression_status_call_raw(threshold,pseudobulk)
}

expression_status_call_raw <- function(threshold, pseudobulk) {

  expression <- pseudobulk > cutoff.int
  pvalue     <- compute_pvalue(pseudobulk, threshold)
  if (is.na(expression)) {
    return('unexpressed')
  }
  status <- if(expression) { 'expressed' } else { 'unexpressed' }
  status <- if(pvalue > 5E-03) { 'unpredictable' } else { status }

  return(status)
}

is_expressed <- function(pseudobulks) {
  cells = names(pseudobulks)
  expressed = sapply(cells, function(cell) { expression_status_call( attributes(alldata_pseudobulk)$thresholds[cell], pseudobulks[cell] ) })
  ifelse(expressed == 'expressed', TRUE, ifelse(expressed == 'unexpressed',FALSE,NA))
}
```

# generating pseudobulkpresence dataset for tabula sapiens

```{r human_load}

alldata_pseudobulk = readRDS('alldata_human.Rds')
alldata_pseudopresence = readRDS('alldata_human_pseudopresence.Rds')

```

```{r get_human_attr}

attributes(alldata_pseudobulk)$thresholds = sapply(attributes(alldata_pseudobulk)$clustersize, compute_threshold)

```

```{r relabel_human1}

alldata_pseudobulk_clean = alldata_pseudobulk
alldata_pseudopresence_clean = alldata_pseudopresence

```

```{r relabel_human2}

colnames(alldata_pseudobulk_clean) = dplyr::recode(colnames(alldata_pseudobulk_clean), TMEM5='RXYLT1',TMEM246='PGAP4',KDELC1='POGLUT2',KDELC2='POGLUT3')
colnames(alldata_pseudopresence_clean) = dplyr::recode(colnames(alldata_pseudopresence_clean), TMEM5='RXYLT1',TMEM246='PGAP4',KDELC1='POGLUT2',KDELC2='POGLUT3')

```

```{r filter_human1}

required_genes=c('DPAGT1','ALG2')
passing_cells = apply(apply(alldata_pseudobulk_clean[,required_genes],2,is_expressed) ,1,all)
passing_cells = passing_cells & apply(apply(alldata_pseudobulk_clean[,c('ALG13','ALG14')],2,is_expressed),1,any)

```

```{r filter_human2}

alldata_pseudobulk_clean = alldata_pseudobulk_clean[which(passing_cells),]
alldata_pseudopresence_clean = alldata_pseudopresence_clean[which(passing_cells),]

```

```{r filter_human3}

attributes(alldata_pseudobulk_clean)$clustersize <- attributes(alldata_pseudobulk)$clustersize[which(passing_cells)]
attributes(alldata_pseudobulk_clean)$thresholds <- attributes(alldata_pseudobulk)$thresholds[which(passing_cells)]

```

```{r remove_human}

remove(alldata_pseudobulk)
remove(alldata_pseudopresence)

```

```{r merge_column_human}

cluster_threshold_cellnum.df <- cbind.data.frame(
  cluster = rep(rownames(alldata_pseudobulk_clean), each = ncol(alldata_pseudobulk_clean)),
  threshold = rep(attributes(alldata_pseudobulk_clean)$thresholds, each = ncol(alldata_pseudobulk_clean)),
  cellnum = rep(attributes(alldata_pseudobulk_clean)$clustersize, each = ncol(alldata_pseudobulk_clean)))

```

```{r pseudobulk_prepare}

alldata_pseudobulk_clean_melt <- reshape2:::melt.matrix(t(alldata_pseudobulk_clean), varnames = c('gene','cluster'), value.name = 'pseudobulk')
alldata_pseudobulk_clean_melt <- within(alldata_pseudobulk_clean_melt, {
  cluster = as.character(cluster)
  gene = as.character(gene) })
remove(alldata_pseudobulk_clean)

```

```{r pseudopresence_prepare}

alldata_pseudopresence_clean_melt <- reshape2:::melt.matrix(t(alldata_pseudopresence_clean), varnames = c('gene','cluster'), value.name = 'pseudopresence')
alldata_pseudopresence_clean_melt <- within(alldata_pseudopresence_clean_melt, {
  cluster = as.character(cluster)
  gene = as.character(gene) })
remove(alldata_pseudopresence_clean)

```

```{r merge_all_human}

alldata_clean_melt <-
  if(all(cluster_threshold_cellnum.df$cluster == alldata_pseudobulk_clean_melt$cluster) &
     all(gsub('SMALL_', '', alldata_pseudobulk_clean_melt$cluster) == alldata_pseudopresence_clean_melt$cluster) &
     all(alldata_pseudobulk_clean_melt$gene == alldata_pseudopresence_clean_melt$gene)) {
    cbind.data.frame(
      alldata_pseudobulk_clean_melt,
      pseudopresence = alldata_pseudopresence_clean_melt$pseudopresence,
      cluster_threshold_cellnum.df[,c('threshold','cellnum')])
  } else { stop ('rows are out of order') }

remove(cluster_threshold_cellnum.df)
remove(alldata_pseudobulk_clean_melt)
remove(alldata_pseudopresence_clean_melt)

```

```{r add_variables_human}

alldata_clean_melt$orig.NA <- is.na(alldata_clean_melt$pseudobulk)
alldata_clean_melt$pseudobulk[is.na(alldata_clean_melt$pseudobulk)] <- 0
alldata_clean_melt$pvalue <- with(alldata_clean_melt, compute_pvalue(pseudobulk, threshold))
alldata_clean_melt$`status5E-03` <- with(
  alldata_clean_melt,
  addNA(factor(
    ifelse(
      pvalue > 5E-03, 'unpredictable',
      ifelse(pseudobulk > cutoff.int, 'expressed', 'unexpressed')),
    levels = rev(c('unexpressed','unpredictable','expressed')))))

```

```{r predict_human}
# computing gene ubiquitous status via pseudopresence model ----
predicted_percent_pseudopresence <-
  compute_sample_pop(
    alldata_clean_melt$pseudobulk,
    pseudopresence_model.lm$glycogene_mean_fit.lm,
    pseudopresence_model.lm$glycogene_sd_fit.lm)

alldata_clean_melt$pseudopresence.pred_mu <- (predicted_percent_pseudopresence$mu/100) * alldata_clean_melt$cellnum
alldata_clean_melt$pseudopresence.pred_sd <- (predicted_percent_pseudopresence$s/100) * alldata_clean_melt$cellnum
alldata_clean_melt$pseudopresence.pval <-
  compute_pseudopresence_pvalue(
    alldata_clean_melt$pseudopresence,
    alldata_clean_melt$pseudopresence.pred_mu,
    alldata_clean_melt$pseudopresence.pred_sd,
    nrow(pseudopresence_model.lm$glycogene_sd_fit.lm$model))

```

```{r save_human}

attributes(alldata_clean_melt)$cutoff.int <- cutoff.int
saveRDS(alldata_clean_melt, 'alldata_pseudobulkpresence.Rds')

```

# generating pseudobulkpresence dataset for panglao musculus

```{r mouse_genes}

ortho_trans_table=with(read.delim('mouse_orthology_human_genes.tsv'),setNames(mouse_symbol,human_symbol))
glycogenes=read.delim('glycogenes.tsv',header=F)$V1

```

```{r mouse_load}

alldata_pseudobulk = readRDS('alldata_mouse.Rds')

```

```{r get_mouse_attr}

attributes(alldata_pseudobulk)$thresholds = sapply(attributes(alldata_pseudobulk)$clustersize, compute_threshold)

```

```{r relabel_mouse1}

alldata_pseudobulk_clean = alldata_pseudobulk

```

```{r filter_mouse1}

required_genes=c('Dpagt1','Alg2')
passing_cells = apply(apply(alldata_pseudobulk_clean[,required_genes],2,is_expressed) ,1,all)
passing_cells = passing_cells & apply(apply(alldata_pseudobulk_clean[,c('Alg13','Alg14')],2,is_expressed),1,any)

```

```{r filter_mouse2}

alldata_pseudobulk_clean = alldata_pseudobulk_clean[which(passing_cells),]

```

```{r filter_mouse3}

colnames(alldata_pseudobulk_clean) = dplyr::recode(colnames(alldata_pseudobulk_clean), Wbscr17='Galnt17', Tmem5='Rxylt1', Tmem246='Pgap4', Kdelc1='Poglut2', Kdelc2='Poglut3')

```

```{r filter_mouse4}

alldata_pseudobulk_clean = alldata_pseudobulk_clean[,c(intersect(ortho_trans_table[glycogenes], colnames(alldata_pseudobulk_clean)),'Dag1')]

```

```{r filter_mouse5}

colnames(alldata_pseudobulk_clean) = dplyr::recode(colnames(alldata_pseudobulk_clean), Large='Large1', Gyltl1b='Large2')

```

```{r filter_mouse6}

attributes(alldata_pseudobulk_clean)$clustersize <- attributes(alldata_pseudobulk)$clustersize[which(passing_cells)]
attributes(alldata_pseudobulk_clean)$thresholds <- attributes(alldata_pseudobulk)$thresholds[which(passing_cells)]

```

```{r remove_mouse}

remove(alldata_pseudobulk)

```

```{r merge_column_mouse}

cluster_threshold_cellnum.df <- cbind.data.frame(
  cluster = rep(rownames(alldata_pseudobulk_clean), each = ncol(alldata_pseudobulk_clean)),
  threshold = rep(attributes(alldata_pseudobulk_clean)$thresholds, each = ncol(alldata_pseudobulk_clean)),
  cellnum = rep(attributes(alldata_pseudobulk_clean)$clustersize, each = ncol(alldata_pseudobulk_clean)))

```

```{r pseudobulk_prepare_mouse}

alldata_pseudobulk_clean_melt <- reshape2:::melt.matrix(t(alldata_pseudobulk_clean), varnames = c('gene','cluster'), value.name = 'pseudobulk')
alldata_pseudobulk_clean_melt <- within(alldata_pseudobulk_clean_melt, {
  cluster = as.character(cluster)
  gene = as.character(gene) })
remove(alldata_pseudobulk_clean)

```

```{r merge_all_mouse}

alldata_clean_melt <-
  if(all(cluster_threshold_cellnum.df$cluster == alldata_pseudobulk_clean_melt$cluster)) {
    cbind.data.frame(
      alldata_pseudobulk_clean_melt,
      cluster_threshold_cellnum.df[,c('threshold','cellnum')])
  } else { stop ('rows are out of order') }

remove(cluster_threshold_cellnum.df)
remove(alldata_pseudobulk_clean_melt)

```

```{r add_variables_mouse}

alldata_clean_melt$orig.NA <- is.na(alldata_clean_melt$pseudobulk)
alldata_clean_melt$pseudobulk[is.na(alldata_clean_melt$pseudobulk)] <- 0
alldata_clean_melt$pvalue <- with(alldata_clean_melt, compute_pvalue(pseudobulk, threshold))
alldata_clean_melt$`status5E-03` <- with(
  alldata_clean_melt,
  addNA(factor(
    ifelse(
      pvalue > 5E-03, 'unpredictable',
      ifelse(pseudobulk > cutoff.int, 'expressed', 'unexpressed')),
    levels = rev(c('unexpressed','unpredictable','expressed')))))

```

```{r predict_mouse}
# computing gene ubiquitous status via pseudopresence model ----
predicted_percent_pseudopresence <-
  compute_sample_pop(
    alldata_clean_melt$pseudobulk,
    pseudopresence_model.lm$glycogene_mean_fit.lm,
    pseudopresence_model.lm$glycogene_sd_fit.lm)

alldata_clean_melt$pseudopresence.pred_mu <- (predicted_percent_pseudopresence$mu/100) * alldata_clean_melt$cellnum
alldata_clean_melt$pseudopresence.pred_sd <- (predicted_percent_pseudopresence$s/100) * alldata_clean_melt$cellnum

```

```{r save_mouse}

attributes(alldata_clean_melt)$cutoff.int <- cutoff.int
saveRDS(alldata_clean_melt, 'glycogene_mouse_panglaodb_pseudobulkpresence.Rds')
remove(alldata_clean_melt)

```
