==> mount/preprocess/generate_single_cell_data/2.generate_log_seg_norm.Rmd <==
# Log-SEG-Normalised data generation
# loading libraries

```{r libraries}
library(dplyr)
library(compositions)
```

# loading pseudobulk model

```{r pseudobulk_model1}
expression_model.lm <- readRDS('lm_chromium10x_mupb_norm_nziqr_threshold_cutoff.Rds')
cutoff.int <- mean(expression_model.lm$xy.df$y[1:7])

compute_threshold <- function(ncells) {

  s <-
    stats::predict(
      expression_model.lm$model2.lm,
      newdata = data.frame(x = ncells),
      se.fit = TRUE,
      level = 0.95,
      interval = "confidence")$fit[,'fit']

  return(s)

}
```

```{r pseudobulk_model2}
compute_pvalue <- function(pseudobulk, threshold) {

  n   <- 7 # chosen based on the number of pseudobulk sizes that model is fit to
  ts  <- (cutoff.int - pseudobulk)/(threshold/sqrt(n))
  p   <- 2*pt(-abs(ts), n-1)

  return(p)

}

expression_status_call <- function(threshold,pseudobulk) {
  if (is.na(pseudobulk)) {
    return("unexpressed")
  }
  if (pseudobulk == 0) {
    return("unexpressed")
  }

  expression_status_call_raw(threshold,pseudobulk)
}

expression_status_call_raw <- function(threshold, pseudobulk) {

  expression <- pseudobulk > cutoff.int
  pvalue     <- compute_pvalue(pseudobulk, threshold)
  if (is.na(expression)) {
    return('unexpressed')
  }
  status <- if(expression) { 'expressed' } else { 'unexpressed' }
  status <- if(pvalue > 5E-03) { 'unpredictable' } else { status }

  return(status)
}

is_expressed <- function(pseudobulks) {
  cells = names(pseudobulks)
  expressed = sapply(cells, function(cell) { expression_status_call( attributes(alldata)$thresholds[cell], pseudobulks[cell] ) })
  ifelse(expressed == 'expressed', TRUE, ifelse(expressed == 'unexpressed',FALSE,NA))
}
```

# loading housekeeping genes

```{r load_housekeeping}
segs=unique(names(readRDS('seg_lookup.Rds')))
ortho_trans_table=with(read.delim('mouse_orthology_human_genes.tsv'),setNames(mouse_symbol,human_symbol))
glycogenes=unique(c(read.delim('glycogenes.tsv',header=F)$V1, segs))
```

# preparing mouse data

```{r load_datasets_mouse}
alldata = readRDS('alldata_mouse.Rds')
alldata_clean_mouse = alldata[!grepl('SMALL',rownames(alldata)),]
attributes(alldata_clean_mouse)$clustersize = attributes(alldata)$clustersize
colnames(alldata_clean_mouse) = recode(colnames(alldata_clean_mouse), Tmem5='Rxylt1',Tmem246='Pgap4',Kdelc1='Poglut2',Kdelc2='Poglut3')
rm(alldata)
```

```{r retreive_attr_mouse}
alldata=list()
attributes(alldata)$clustersize = c(attributes(alldata_clean_mouse)$clustersize)
attributes(alldata)$thresholds = sapply(attributes(alldata)$clustersize,compute_threshold)
```

```{r filter_mouse}
required_genes=c('Dpagt1','Alg2')
passing_cells = apply(apply(alldata_clean_mouse[,required_genes],2,is_expressed) ,1,all)
passing_cells = passing_cells & apply(apply(alldata_clean_mouse[,c('Alg13','Alg14')],2,is_expressed),1,any)
alldata_clean_mouse = alldata_clean_mouse[which(passing_cells),]
```

## log transform

```{r log_mouse}
alldata_clean_mouse_log = apply(alldata_clean_mouse,2,log)
alldata_clean_mouse_log[!is.finite(alldata_clean_mouse_log)] <- 0
mouse_melt = subset(reshape2::melt(alldata_clean_mouse_log[,intersect(ortho_trans_table[glycogenes],colnames(alldata_clean_mouse_log))]),value != 0)
mouse_melt$Var2 = setNames(names(ortho_trans_table),ortho_trans_table)[as.character(mouse_melt$Var2)]
mouse_melt = cbind(subset(mouse_melt,Var2 != 'cell'),data='mouse')
rm(alldata_clean_mouse)
rm(alldata_clean_mouse_log)
```

```{r merge_mouse}
mouse_melt = merge(
  mouse_melt,
  data.frame(
    Var1 = names(attributes(alldata)$clustersize),
    samples = attributes(alldata)$clustersize))
```

# preparing human data

```{r load_datasets_human}
alldata = readRDS('alldata_human.Rds')
alldata_clean_human = alldata[!grepl('SMALL',rownames(alldata)),]
attributes(alldata_clean_human)$clustersize = attributes(alldata)$clustersize
colnames(alldata_clean_human) = recode(colnames(alldata_clean_human), TMEM5='RXYLT1',TMEM246='PGAP4',KDELC1='POGLUT2',KDELC2='POGLUT3')
rm(alldata)
```

```{r retrieve_attr_human}
alldata=list()
attributes(alldata)$clustersize = c(attributes(alldata_clean_human)$clustersize)
attributes(alldata)$thresholds = sapply(attributes(alldata)$clustersize,compute_threshold)
```

```{r filter_human}
required_genes=c('DPAGT1','ALG2')
passing_cells = apply(apply(alldata_clean_human[,required_genes],2,is_expressed) ,1,all)
passing_cells = passing_cells & apply(apply(alldata_clean_human[,c('ALG13','ALG14')],2,is_expressed),1,any)
alldata_clean_human = alldata_clean_human[which(passing_cells),]
```

## log transform

```{r log_human}
alldata_clean_human_log = apply(alldata_clean_human,2,log)
alldata_clean_human_log[!is.finite(alldata_clean_human_log)] <- 0
human_melt = cbind(subset(reshape2::melt(alldata_clean_human_log[,intersect(glycogenes,colnames(alldata_clean_human_log))]),value != 0),data='human')
human_melt$Var2 = as.character(human_melt$Var2)
rm(alldata_clean_human)
rm(alldata_clean_human_log)
```

```{r merge_human}
human_melt = merge(
  human_melt,
  data.frame(
    Var1 = names(attributes(alldata)$clustersize),
    samples = attributes(alldata)$clustersize))
```

# merging human and mouse data

```{r merge_both}
human_melt$tissue_cell = gsub('_cluster.*$', '', human_melt$Var1)
mouse_melt$tissue_cell = stringr::str_replace(stringr::str_replace(mouse_melt$Var1,'^[^:]+:',''),'_cluster.*','')
full_log = rbind(human_melt,mouse_melt)
full_log$cell = stringr::str_replace( full_log$tissue_cell,'.*:','')
rm(human_melt)
rm(mouse_melt)
```

# normalising by housekeeping genes

```{r retrieve_ref}
ref_data = full_log[full_log$Var2 %in% segs,]
```

```{r summarize_ref}
ref_data = aggregate.data.frame(
  x = list(seg_avg = ref_data$value),
  by = list(Var1 = ref_data$Var1),
  FUN = function(x) { mean(x) })
```

```{r apply_ref}
log_seg_avg_norm = merge(ref_data,full_log,by='Var1')
log_seg_avg_norm$value = log_seg_avg_norm$value - log_seg_avg_norm$seg_avg
clusters <- log_seg_avg_norm$Var1
log_seg_avg_norm = setNames(log_seg_avg_norm[,c('value','tissue_cell','Var2','data','samples')],c('rel_diff','tissue','gene','source','samples'))
attributes(log_seg_avg_norm)$cutoffs = with(ref_data, setNames(log(cutoff.int)-seg_avg, Var1))
attributes(log_seg_avg_norm)$cluster = clusters
saveRDS(log_seg_avg_norm,'pseudobulk_log_seg_norm.Rds')
```
